Recompile extension modules when testing
Adds

    CYTHONIZE=1 $(VENV_PYTHON) setup.py build_ext --inplace --verbose

to the `Makefile`.

This requires that cython is in the dev-requirements: Having it only in
`pyproject.toml` does not work! This is because pip has "isolated
builds", where the requirements listed in `pyproject.toml` are installed
only into a temporary build environment, but not into the environment
from which `pip` is run.